Fix string formatting in clog functions and add functionality to remove templates in color messages
The commit message describes the changes made to two files. The first file is clog.go, which had some issues with string formatting in the DebugF, WarnF, and ErrorF functions. The input parameter is a variadic argument, and it was not properly unpacked using the ellipsis. The commit fixes this issue by changing input to input... in the fmt.Sprintf function calls.

The second file is color.go, which had a new function added called RemoveTemplate. This function removes templates from color messages. Templates are denoted by two curly braces with a pipe character in the middle. For example, {{red|this text is red}}. The function uses regular expressions to find all occurrences of templates and replaces them with the text to the left of the pipe character. The function also calls the existing Template function to remove any nested templates in the message before removing templates.